build: consume io.github.kyrylr lwk 0.16.1

+++ common/src/commonMain/kotlin/com/blockstream/common/gdk/data/Transaction.kt
@@ -263,7 +263,7 @@ data class Transaction constructor(
             it.hasUnblindingData()
         }.map {
             InputUnblindedData(
-                vin = it.ptIdx?.toUInt() ?: 0.toUInt(),
+                vin = it.ptIdx?.toUInt() ?: 0u,
                 assetId = it.assetId ?: "",
                 assetblinder = it.assetblinder ?: "",
                 satoshi = it.satoshi ?: 0,
@@ -275,7 +275,7 @@ data class Transaction constructor(
             it.hasUnblindingData()
         }.map {
             OutputUnblindedData(
-                vout = it.ptIdx?.toUInt() ?: 0.toUInt(),
+                vout = it.ptIdx?.toUInt() ?: 0u,
                 assetId = it.assetId ?: "",
                 assetblinder = it.assetblinder ?: "",
                 satoshi = it.satoshi ?: 0,
